requirements.txt
# 기본 라이브러리
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
albumentations==1.3.1
opencv-python==4.8.1.78

# 데이터 처리
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# 시각화 및 로깅
matplotlib==3.7.2
seaborn==0.12.2
tensorboard>=2.13.0

# 온라인 로깅 (선택적)
wandb>=0.15.0

# 유틸리티
tqdm==4.66.1
PyYAML==6.0.1
Pillow==10.0.0 